refactor(backend): split storage.rs by entity (phase 2)
Decompose 988-line storage.rs into storage/ with one module
per concern:

  types.rs              QueryHistoryEntry, SavedQuery,
                        ConnectionHistoryEntry, SnippetFolder
  serialize.rs          Stored* helpers + DatabaseInfo codec
  migrator.rs           Migrator struct + migrate()
  mod.rs                Storage struct, new(), get_sqlite_connection()
  connections.rs        save/update/get/remove connection + update_last_connected
  queries.rs            save/get query_history + save/get/delete saved_queries
  connection_history.rs save/get connection_history
  settings.rs           get_setting / set_setting
  snippet_folders.rs    CRUD for snippet_folders

All public re-exports (Storage, QueryHistoryEntry, SavedQuery,
ConnectionHistoryEntry, SnippetFolder) preserved through mod.rs
so every existing crate::storage::X path resolves unchanged.

No SQL change. No migration bump. Behavior unchanged.

Part of docs/backend-refactor-plan.md.
