--- apps/desktop/src-tauri/src/storage.rs
@@ -1,988 +1,0 @@
-use std::path::PathBuf;
-use std::sync::Mutex;
-
-use anyhow::Context;
-use rusqlite::{types::Type, Connection};
-use serde::{Deserialize, Serialize};
-use uuid::Uuid;
-
-// Gotta match the IDs in the DB
-const DB_TYPE_POSTGRES: i32 = 1;
-const DB_TYPE_SQLITE: i32 = 2;
-const DB_TYPE_LIBSQL: i32 = 3;
-const DB_TYPE_MYSQL: i32 = 4;
-
-use crate::{database::types::ConnectionInfo, Result};
-
-#[derive(Debug, Serialize, Deserialize)]
-struct StoredPostgresConnection {
-    connection_string: String,
-    #[serde(default)]
-    ssh_config: Option<crate::database::types::SshConfig>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct StoredLibsqlConnection {
-    url: String,
-    auth_token: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct StoredMysqlConnection {
-    connection_string: String,
-    #[serde(default)]
-    ssh_config: Option<crate::database::types::SshConfig>,
-}
-
-fn serialize_connection_data(database_type: &crate::database::types::DatabaseInfo) -> Result<(i32, String)> {
-    match database_type {
-        crate::database::types::DatabaseInfo::Postgres {
-            connection_string,
-            ssh_config,
-        } => Ok((
-            DB_TYPE_POSTGRES,
-            serde_json::to_string(&StoredPostgresConnection {
-                connection_string: connection_string.clone(),
-                ssh_config: ssh_config.clone(),
-            })
-            .context("Failed to serialize Postgres connection data")?,
-        )),
-        crate::database::types::DatabaseInfo::SQLite { db_path } => Ok((DB_TYPE_SQLITE, db_path.clone())),
-        crate::database::types::DatabaseInfo::LibSQL { url, auth_token } => Ok((
-            DB_TYPE_LIBSQL,
-            serde_json::to_string(&StoredLibsqlConnection {
-                url: url.clone(),
-                auth_token: auth_token.clone(),
-            })
-            .context("Failed to serialize LibSQL connection data")?,
-        )),
-        crate::database::types::DatabaseInfo::MySQL {
-            connection_string,
-            ssh_config,
-        } => Ok((
-            DB_TYPE_MYSQL,
-            serde_json::to_string(&StoredMysqlConnection {
-                connection_string: connection_string.clone(),
-                ssh_config: ssh_config.clone(),
-            })
-            .context("Failed to serialize MySQL connection data")?,
-        )),
-    }
-}
-
-fn db_type_from_id(db_type_id: i32) -> &'static str {
-    match db_type_id {
-        DB_TYPE_POSTGRES => "postgres",
-        DB_TYPE_SQLITE => "sqlite",
-        DB_TYPE_LIBSQL => "libsql",
-        DB_TYPE_MYSQL => "mysql",
-        _ => "postgres",
-    }
-}
-
-fn deserialize_database_info(
-    db_type: &str,
-    db_type_id: i32,
-    connection_data: String,
-) -> crate::database::types::DatabaseInfo {
-    let db_type = if db_type.is_empty() {
-        db_type_from_id(db_type_id)
-    } else {
-        db_type
-    };
-
-    match db_type {
-        "postgres" => {
-            if let Ok(stored) = serde_json::from_str::<StoredPostgresConnection>(&connection_data) {
-                crate::database::types::DatabaseInfo::Postgres {
-                    connection_string: stored.connection_string,
-                    ssh_config: stored.ssh_config,
-                }
-            } else {
-                crate::database::types::DatabaseInfo::Postgres {
-                    connection_string: connection_data,
-                    ssh_config: None,
-                }
-            }
-        }
-        "sqlite" => crate::database::types::DatabaseInfo::SQLite {
-            db_path: connection_data,
-        },
-        "libsql" => {
-            if let Ok(stored) = serde_json::from_str::<StoredLibsqlConnection>(&connection_data) {
-                crate::database::types::DatabaseInfo::LibSQL {
-                    url: stored.url,
-                    auth_token: stored.auth_token,
-                }
-            } else {
-                crate::database::types::DatabaseInfo::LibSQL {
-                    url: connection_data,
-                    auth_token: None,
-                }
-            }
-        }
-        "mysql" => {
-            if let Ok(stored) = serde_json::from_str::<StoredMysqlConnection>(&connection_data) {
-                crate::database::types::DatabaseInfo::MySQL {
-                    connection_string: stored.connection_string,
-                    ssh_config: stored.ssh_config,
-                }
-            } else {
-                crate::database::types::DatabaseInfo::MySQL {
-                    connection_string: connection_data,
-                    ssh_config: None,
-                }
-            }
-        }
-        _ => crate::database::types::DatabaseInfo::Postgres {
-            connection_string: connection_data,
-            ssh_config: None,
-        },
-    }
-}
-
-struct Migrator {
-    migrations: &'static [&'static str],
-}
-
-impl Migrator {
-    fn new() -> Self {
-        Self {
-            migrations: &[
-                include_str!("../migrations/001.sql"),
-                include_str!("../migrations/002.sql"),
-                include_str!("../migrations/003.sql"),
-                include_str!("../migrations/004.sql"),
-                include_str!("../migrations/005.sql"),
-                include_str!("../migrations/006.sql"),
-                include_str!("../migrations/007.sql"),
-            ],
-        }
-    }
-
-    fn migrate(&self, conn: &mut Connection) -> anyhow::Result<()> {
-        let current_version: i32 = conn
-            .pragma_query_value(None, "user_version", |row| row.get(0))
-            .context("Failed to get current database version")?;
-
-        let target_version = self.migrations.len() as i32;
-
-        if current_version == target_version {
-            return Ok(());
-        }
-
-        if current_version > target_version {
-            anyhow::bail!(
-                "Database version ({}) is newer than application version ({}). Please update the application.",
-                current_version,
-                target_version
-            );
-        }
-
-        let tx = conn
-            .transaction()
-            .context("Failed to start migration transaction")?;
-
-        for (i, migration) in self.migrations.iter().enumerate() {
-            let migration_version = (i + 1) as i32;
-
-            if migration_version <= current_version {
-                continue;
-            }
-
-            tx.execute_batch(migration).map_err(|err| {
-                anyhow::anyhow!("Failed to execute migration {migration_version}: {err}")
-            })?;
-
-            tx.pragma_update(None, "user_version", migration_version)
-                .with_context(|| format!("Failed to update version to {}", migration_version))?;
-        }
-
-        let integrity_check: String = tx
-            .pragma_query_value(None, "integrity_check", |row| row.get(0))
-            .context("Failed to check database integrity")?;
-
-        anyhow::ensure!(
-            integrity_check == "ok",
-            "Database integrity check failed: {}",
-            integrity_check
-        );
-
-        tx.commit()
-            .context("Failed to commit migration transaction")?;
-
-        conn.execute("PRAGMA optimize", [])
-            .context("Failed to optimize database")?;
-
-        Ok(())
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, specta::Type)]
-pub struct QueryHistoryEntry {
-    pub id: i64,
-    pub connection_id: String,
-    pub query_text: String,
-    pub executed_at: i64,
-    pub duration_ms: Option<i64>,
-    pub status: String,
-    pub row_count: i64,
-    pub error_message: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize, specta::Type)]
-pub struct SavedQuery {
-    pub id: i64,
-    pub name: String,
-    pub description: Option<String>,
-    pub query_text: String,
-    pub connection_id: Option<Uuid>,
-    pub tags: Option<String>,
-    pub category: Option<String>,
-    pub created_at: i64,
-    pub updated_at: i64,
-    pub favorite: bool,
-    pub is_snippet: bool,
-    pub is_system: bool,
-    pub language: Option<String>,
-    pub folder_id: Option<i64>,
-}
-
-
-
-#[derive(Debug, Serialize, Deserialize, specta::Type)]
-pub struct ConnectionHistoryEntry {
-    pub id: i64,
-    pub connection_id: String,
-    pub connection_name: String,
-    pub database_type: String,
-    pub attempted_at: i64,
-    pub success: bool,
-    pub error_message: Option<String>,
-    pub duration_ms: Option<i64>,
-}
-
-#[derive(Debug, Serialize, Deserialize, specta::Type)]
-pub struct SnippetFolder {
-    pub id: i64,
-    pub name: String,
-    pub parent_id: Option<i64>,
-    pub color: Option<String>,
-    pub created_at: i64,
-    pub updated_at: i64,
-}
-
-#[derive(Debug)]
-pub struct Storage {
-    conn: Mutex<Connection>,
-}
-
-impl Storage {
-    pub fn new(db_path: PathBuf) -> Result<Self> {
-        if let Some(parent) = db_path.parent() {
-            std::fs::create_dir_all(parent).with_context(|| {
-                format!("Failed to create database directory: {}", parent.display())
-            })?;
-        }
-
-        let mut conn = Connection::open(&db_path)
-            .with_context(|| format!("Failed to open SQLite database at {}", db_path.display()))?;
-
-        conn.execute_batch(
-            "
-            PRAGMA foreign_keys = ON;
-            PRAGMA journal_mode = WAL;
-            PRAGMA synchronous = FULL;
-            PRAGMA busy_timeout = 30000;
-            PRAGMA case_sensitive_like = ON;
-            PRAGMA extended_result_codes = ON;
-            ",
-        )
-        .context("Failed to execute database initialization SQL")?;
-
-        let migrator = Migrator::new();
-        migrator.migrate(&mut conn)?;
-
-        Ok(Self {
-            conn: Mutex::new(conn),
-        })
-    }
-
-    /// Get a reference to the underlying connection (locked)
-    pub fn get_sqlite_connection(&self) -> Result<std::sync::MutexGuard<'_, Connection>> {
-        self.conn
-            .lock()
-            .map_err(|e| crate::Error::Any(anyhow::anyhow!("Failed to lock connection: {}", e)))
-    }
-    pub fn save_connection(&self, connection: &ConnectionInfo) -> Result<()> {
-        let now = chrono::Utc::now().timestamp();
-        let conn = self.conn.lock().unwrap();
-
-        let (db_type_id, connection_data) = serialize_connection_data(&connection.database_type)?;
-
-        let encrypted_data = crate::security::encrypt(&connection_data)
-            .context("Failed to encrypt connection data")?;
-
-        conn.execute(
-            "INSERT OR REPLACE INTO connections 
-             (id, name, connection_data, database_type_id, created_at, updated_at, sort_order, color) 
-             VALUES (?1, ?2, ?3, ?4, ?5, ?6, 
-                (SELECT COALESCE(MAX(sort_order), 0) + 1 FROM connections), ?7)",
-            (
-                &connection.id.to_string(),
-                &connection.name,
-                &encrypted_data,
-                db_type_id,
-                now,
-                now,
-                connection.color.as_deref(),
-            ),
-        )
-        .context("Failed to save connection")?;
-
-        Ok(())
-    }
-
-    pub fn update_connection(&self, connection: &ConnectionInfo) -> Result<()> {
-        let now = chrono::Utc::now().timestamp();
-        let conn = self.conn.lock().unwrap();
-
-        let (db_type_id, connection_data) = serialize_connection_data(&connection.database_type)?;
-
-        let encrypted_data = crate::security::encrypt(&connection_data)
-            .context("Failed to encrypt connection data")?;
-
-        let updated_rows = conn
-            .execute(
-                "UPDATE connections 
-             SET name = ?2, connection_data = ?3, database_type_id = ?4, updated_at = ?5, color = ?6
-             WHERE id = ?1",
-                (
-                    &connection.id.to_string(),
-                    &connection.name,
-                    &encrypted_data,
-                    db_type_id,
-                    now,
-                    connection.color.as_deref(),
-                ),
-            )
-            .context("Failed to update connection")?;
-
-        if updated_rows == 0 {
-            return Err(crate::Error::Any(anyhow::anyhow!(
-                "Connection not found: {}",
-                connection.id
-            )));
-        }
-
-        Ok(())
-    }
-
-    pub fn get_connection(&self, connection_id: &Uuid) -> Result<Option<ConnectionInfo>> {
-        let conn = self.conn.lock().unwrap();
-        let mut stmt = conn
-            .prepare(
-                "SELECT c.id, c.name, c.connection_data, 
-                        c.database_type_id,
-                        COALESCE(dt.name, 'postgres') as db_type,
-                        c.last_connected_at, c.favorite, c.color, c.sort_order
-                 FROM connections c
-                 LEFT JOIN database_types dt ON c.database_type_id = dt.id
-                 WHERE c.id = ?1",
-            )
-            .context("Failed to prepare statement")?;
-
-        let mut rows = stmt
-            .query_map([connection_id.to_string()], |row| {
-                let raw_data: String = row.get(2)?;
-                let db_type_id: i32 = row.get(3)?;
-                let db_type: String = row.get(4)?;
-
-                // Try to decrypt, otherwise assume plaintext (lazy migration)
-                let connection_data = match crate::security::decrypt(&raw_data) {
-                    Ok(decrypted) => decrypted,
-                    Err(_) => raw_data, // Fallback to plaintext
-                };
-
-                let database_type = deserialize_database_info(&db_type, db_type_id, connection_data);
-
-                Ok(ConnectionInfo {
-                    id: {
-                        let id: String = row.get(0)?;
-                        Uuid::parse_str(&id).map_err(|err| {
-                            rusqlite::Error::FromSqlConversionFailure(0, Type::Text, Box::new(err))
-                        })?
-                    },
-                    name: row.get(1)?,
-                    database_type,
-                    connected: false,
-                    last_connected_at: row.get(5).ok(),
-                    favorite: row.get(6).ok(),
-                    color: row.get(7).ok(),
-                    sort_order: row.get(8).ok(),
-                    created_at: None,
-                    updated_at: None,
-                    pin_hash: None,
-                })
-            })
-            .context("Failed to query connection")?;
-
-        match rows.next() {
-            Some(row) => Ok(Some(row.map_err(|e| {
-                anyhow::anyhow!("Failed to process connection row: {}", e)
-            })?)),
-            None => Ok(None),
-        }
-    }
-
-    pub fn get_connections(&self) -> Result<Vec<ConnectionInfo>> {
-        let conn = self.conn.lock().unwrap();
-        let mut stmt = conn
-            .prepare(
-                "SELECT c.id, c.name, c.connection_data, 
-                        c.database_type_id,
-                        COALESCE(dt.name, 'postgres') as db_type,
-                        c.last_connected_at, c.favorite, c.color, c.sort_order
-                 FROM connections c
-                 LEFT JOIN database_types dt ON c.database_type_id = dt.id
-                 ORDER BY c.sort_order, c.name",
-            )
-            .context("Failed to prepare statement")?;
-
-        let rows = stmt
-            .query_map([], |row| {
-                let raw_data: String = row.get(2)?;
-                let db_type_id: i32 = row.get(3)?;
-                let db_type: String = row.get(4)?;
-
-                // Try to decrypt, otherwise assume plaintext (lazy migration)
-                let connection_data = match crate::security::decrypt(&raw_data) {
-                    Ok(decrypted) => decrypted,
-                    Err(_) => raw_data, // Fallback to plaintext
-                };
-
-                let database_type = deserialize_database_info(&db_type, db_type_id, connection_data);
-
-                Ok(ConnectionInfo {
-                    id: {
-                        let id: String = row.get(0)?;
-                        Uuid::parse_str(&id).map_err(|err| {
-                            rusqlite::Error::FromSqlConversionFailure(0, Type::Text, Box::new(err))
-                        })?
-                    },
-                    name: row.get(1)?,
-                    database_type,
-                    connected: false,
-                    last_connected_at: row.get(5).ok(),
-                    favorite: row.get(6).ok(),
-                    color: row.get(7).ok(),
-                    sort_order: row.get(8).ok(),
-                    created_at: None,
-                    updated_at: None,
-                    pin_hash: None,
-                })
-            })
-            .context("Failed to query connections")?;
-
-        let mut connections = Vec::new();
-        for row in rows {
-            connections
-                .push(row.map_err(|e| anyhow::anyhow!("Failed to process connection row: {}", e))?);
-        }
-
-        Ok(connections)
-    }
-
-    pub fn remove_connection(&self, connection_id: &Uuid) -> Result<()> {
-        let conn = self.conn.lock().unwrap();
-        conn.execute(
-            "DELETE FROM connections WHERE id = ?1",
-            [connection_id.to_string()],
-        )
-        .context("Failed to remove connection")?;
-        Ok(())
-    }
-
-    pub fn update_last_connected(&self, connection_id: &Uuid) -> Result<()> {
-        let now = chrono::Utc::now().timestamp();
-        let conn = self.conn.lock().unwrap();
-        conn.execute(
-            "UPDATE connections SET last_connected_at = ?1 WHERE id = ?2",
-            (now, connection_id.to_string()),
-        )
-        .context("Failed to update last connected time")?;
-        Ok(())
-    }
-
-    pub fn save_query_history(&self, entry: &QueryHistoryEntry) -> Result<()> {
-        let conn = self.conn.lock().unwrap();
-        conn.execute(
-            "INSERT INTO query_history 
-             (connection_id, query_text, executed_at, duration_ms, status, row_count, error_message)
-             VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)",
-            (
-                &entry.connection_id,
-                &entry.query_text,
-                entry.executed_at,
-                entry.duration_ms,
-                &entry.status,
-                entry.row_count,
-                &entry.error_message,
-            ),
-        )
-        .context("Failed to save query history")?;
-        Ok(())
-    }
-
-    pub fn get_query_history(
-        &self,
-        connection_id: &str,
-        limit: Option<i64>,
-    ) -> Result<Vec<QueryHistoryEntry>> {
-        let limit = limit.unwrap_or(100);
-        let conn = self.conn.lock().unwrap();
-        let mut stmt = conn.prepare(
-            "SELECT id, connection_id, query_text, executed_at, duration_ms, status, row_count, error_message
-             FROM query_history 
-             WHERE connection_id = ?1 
-             ORDER BY executed_at DESC 
-             LIMIT ?2"
-        ).context("Failed to prepare query history statement")?;
-
-        let rows = stmt
-            .query_map((connection_id, limit), |row| {
-                Ok(QueryHistoryEntry {
-                    id: row.get(0)?,
-                    connection_id: row.get(1)?,
-                    query_text: row.get(2)?,
-                    executed_at: row.get(3)?,
-                    duration_ms: row.get(4)?,
-                    status: row.get(5)?,
-                    row_count: row.get(6)?,
-                    error_message: row.get(7)?,
-                })
-            })
-            .context("Failed to query history")?;
-
-        let mut history = Vec::new();
-        for row in rows {
-            history.push(row.context("Failed to process history row")?);
-        }
-
-        Ok(history)
-    }
-
-    pub fn get_setting(&self, key: &str) -> Result<Option<String>> {
-        let conn = self.conn.lock().unwrap();
-        let mut stmt = conn
-            .prepare("SELECT value FROM app_settings WHERE key = ?1")
-            .context("Failed to prepare settings statement")?;
-        let mut rows = stmt
-            .query_map([key], |row| row.get::<_, String>(0))
-            .context("Failed to query settings")?;
-
-        if let Some(row) = rows.next() {
-            Ok(Some(row.context("Failed to get setting value")?))
-        } else {
-            Ok(None)
-        }
-    }
-
-    pub fn set_setting(&self, key: &str, value: &str) -> Result<()> {
-        let now = chrono::Utc::now().timestamp();
-        let conn = self.conn.lock().unwrap();
-        conn.execute(
-            "INSERT OR REPLACE INTO app_settings (key, value, updated_at) VALUES (?1, ?2, ?3)",
-            (key, value, now),
-        )
-        .context("Failed to set setting")?;
-        Ok(())
-    }
-
-    pub fn save_query(&self, query: &SavedQuery) -> Result<i64> {
-        let now = chrono::Utc::now().timestamp();
-        let conn = self.conn.lock().unwrap();
-
-        if query.id == 0 {
-            conn.execute(
-                "INSERT INTO saved_queries 
-                 (name, description, query_text, connection_id, tags, category, created_at, updated_at, favorite, is_snippet, is_system, language, folder_id) 
-                 VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12, ?13)",
-                (
-                    &query.name,
-                    &query.description,
-                    &query.query_text,
-                    query.connection_id.as_ref().map(|u| u.to_string()),
-                    &query.tags,
-                    &query.category,
-                    now,
-                    now,
-                    query.favorite,
-                    query.is_snippet,
-                    query.is_system,
-                    &query.language,
-                    query.folder_id,
-                ),
-
-            ).context("Failed to insert saved query")?;
-            Ok(conn.last_insert_rowid())
-        } else {
-            conn.execute(
-                "UPDATE saved_queries 
-                 SET name = ?1, description = ?2, query_text = ?3, connection_id = ?4, 
-                     tags = ?5, category = ?6, updated_at = ?7, favorite = ?8, is_snippet = ?9, is_system = ?10, language = ?11, folder_id = ?12
-                 WHERE id = ?13",
-                (
-                    &query.name,
-                    &query.description,
-                    &query.query_text,
-                    &query.connection_id.map(|id| id.to_string()),
-                    &query.tags,
-                    &query.category,
-                    now,
-                    query.favorite,
-                    query.is_snippet,
-                    query.is_system,
-                    &query.language,
-                    query.folder_id,
-                    query.id,
-                ),
-
-            )
-            .context("Failed to update saved query")?;
-            Ok(query.id)
-        }
-    }
-
-    pub fn get_saved_queries(&self, connection_id: Option<&Uuid>) -> Result<Vec<SavedQuery>> {
-        let conn = self.conn.lock().unwrap();
-
-        let mut queries = Vec::new();
-
-        if let Some(conn_id) = connection_id {
-            let mut stmt = conn.prepare(
-                "SELECT id, name, description, query_text, connection_id, tags, category, created_at, updated_at, favorite, is_snippet, is_system, language, folder_id
-                 FROM saved_queries 
-                 WHERE connection_id = ?1 OR connection_id IS NULL
-                 ORDER BY favorite DESC, created_at DESC"
-            ).context("Failed to prepare saved queries statement")?;
-
-
-            let rows = stmt
-                .query_map([conn_id.to_string()], |row| {
-                    Ok(SavedQuery {
-                        id: row.get(0)?,
-                        name: row.get(1)?,
-                        description: row.get(2)?,
-                        query_text: row.get(3)?,
-                        connection_id: {
-                            let id: Option<String> = row.get(4)?;
-                            match id {
-                                Some(id) => Some(Uuid::parse_str(&id).map_err(|err| {
-                                    rusqlite::Error::FromSqlConversionFailure(
-                                        0,
-                                        Type::Text,
-                                        Box::new(err),
-                                    )
-                                })?),
-                                None => None,
-                            }
-                        },
-                        tags: row.get(5)?,
-                        category: row.get(6)?,
-                        created_at: row.get(7)?,
-                        updated_at: row.get(8)?,
-                        favorite: row.get(9)?,
-                        is_snippet: row.get(10)?,
-                        is_system: row.get(11)?,
-                        language: row.get(12)?,
-                        folder_id: row.get(13)?,
-                    })
-
-                })
-                .context("Failed to query saved queries")?;
-
-            for row in rows {
-                queries.push(row.context("Failed to process saved query row")?);
-            }
-        } else {
-            let mut stmt = conn.prepare(
-                "SELECT id, name, description, query_text, connection_id, tags, category, created_at, updated_at, favorite, is_snippet, is_system, language, folder_id
-                 FROM saved_queries 
-                 ORDER BY favorite DESC, created_at DESC"
-            ).context("Failed to prepare saved queries statement")?;
-
-
-            let rows = stmt
-                .query_map([], |row| {
-                    Ok(SavedQuery {
-                        id: row.get(0)?,
-                        name: row.get(1)?,
-                        description: row.get(2)?,
-                        query_text: row.get(3)?,
-                        connection_id: {
-                            let id: Option<String> = row.get(4)?;
-                            match id {
-                                Some(id) => Some(Uuid::parse_str(&id).map_err(|err| {
-                                    rusqlite::Error::FromSqlConversionFailure(
-                                        0,
-                                        Type::Text,
-                                        Box::new(err),
-                                    )
-                                })?),
-                                None => None,
-                            }
-                        },
-                        tags: row.get(5)?,
-                        category: row.get(6)?,
-                        created_at: row.get(7)?,
-                        updated_at: row.get(8)?,
-                        favorite: row.get(9)?,
-                        is_snippet: row.get(10)?,
-                        is_system: row.get(11)?,
-                        language: row.get(12)?,
-                        folder_id: row.get(13)?,
-                    })
-
-                })
-                .context("Failed to query saved queries")?;
-
-            for row in rows {
-                queries.push(row.context("Failed to process saved query row")?);
-            }
-        }
-
-        Ok(queries)
-    }
-
-    pub fn delete_saved_query(&self, id: i64) -> Result<()> {
-        let conn = self.conn.lock().unwrap();
-        conn.execute("DELETE FROM saved_queries WHERE id = ?1", [id])
-            .context("Failed to delete saved query")?;
-        Ok(())
-    }
-
-    pub fn save_connection_history(&self, entry: &ConnectionHistoryEntry) -> Result<()> {
-        let conn = self.conn.lock().unwrap();
-        conn.execute(
-            "INSERT INTO connection_history 
-             (connection_id, connection_name, database_type, attempted_at, success, error_message, duration_ms)
-             VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)",
-            (
-                &entry.connection_id,
-                &entry.connection_name,
-                &entry.database_type,
-                entry.attempted_at,
-                entry.success,
-                &entry.error_message,
-                entry.duration_ms,
-            ),
-        )
-        .context("Failed to save connection history")?;
-        Ok(())
-    }
-
-    pub fn get_connection_history(
-        &self,
-        db_type_filter: Option<&str>,
-        success_filter: Option<bool>,
-        limit: Option<i64>,
-    ) -> Result<Vec<ConnectionHistoryEntry>> {
-        let limit = limit.unwrap_or(50);
-        let conn = self.conn.lock().unwrap();
-
-        let _query = match (db_type_filter, success_filter) {
-            (Some(db_type), Some(success)) => {
-                let mut stmt = conn.prepare(
-                    "SELECT id, connection_id, connection_name, database_type, attempted_at, success, error_message, duration_ms
-                     FROM connection_history 
-                     WHERE database_type = ?1 AND success = ?2
-                     ORDER BY attempted_at DESC 
-                     LIMIT ?3"
-                ).context("Failed to prepare connection history statement")?;
-
-                let rows = stmt
-                    .query_map((db_type, success, limit), |row| {
-                        Ok(ConnectionHistoryEntry {
-                            id: row.get(0)?,
-                            connection_id: row.get(1)?,
-                            connection_name: row.get(2)?,
-                            database_type: row.get(3)?,
-                            attempted_at: row.get(4)?,
-                            success: row.get(5)?,
-                            error_message: row.get(6)?,
-                            duration_ms: row.get(7)?,
-                        })
-                    })
-                    .context("Failed to query connection history")?;
-
-                let mut history = Vec::new();
-                for row in rows {
-                    history.push(row.context("Failed to process history row")?);
-                }
-                return Ok(history);
-            }
-            (Some(db_type), None) => {
-                let mut stmt = conn.prepare(
-                    "SELECT id, connection_id, connection_name, database_type, attempted_at, success, error_message, duration_ms
-                     FROM connection_history 
-                     WHERE database_type = ?1
-                     ORDER BY attempted_at DESC 
-                     LIMIT ?2"
-                ).context("Failed to prepare connection history statement")?;
-
-                let rows = stmt
-                    .query_map((db_type, limit), |row| {
-                        Ok(ConnectionHistoryEntry {
-                            id: row.get(0)?,
-                            connection_id: row.get(1)?,
-                            connection_name: row.get(2)?,
-                            database_type: row.get(3)?,
-                            attempted_at: row.get(4)?,
-                            success: row.get(5)?,
-                            error_message: row.get(6)?,
-                            duration_ms: row.get(7)?,
-                        })
-                    })
-                    .context("Failed to query connection history")?;
-
-                let mut history = Vec::new();
-                for row in rows {
-                    history.push(row.context("Failed to process history row")?);
-                }
-                return Ok(history);
-            }
-            (None, Some(success)) => {
-                let mut stmt = conn.prepare(
-                    "SELECT id, connection_id, connection_name, database_type, attempted_at, success, error_message, duration_ms
-                     FROM connection_history 
-                     WHERE success = ?1
-                     ORDER BY attempted_at DESC 
-                     LIMIT ?2"
-                ).context("Failed to prepare connection history statement")?;
-
-                let rows = stmt
-                    .query_map((success, limit), |row| {
-                        Ok(ConnectionHistoryEntry {
-                            id: row.get(0)?,
-                            connection_id: row.get(1)?,
-                            connection_name: row.get(2)?,
-                            database_type: row.get(3)?,
-                            attempted_at: row.get(4)?,
-                            success: row.get(5)?,
-                            error_message: row.get(6)?,
-                            duration_ms: row.get(7)?,
-                        })
-                    })
-                    .context("Failed to query connection history")?;
-
-                let mut history = Vec::new();
-                for row in rows {
-                    history.push(row.context("Failed to process history row")?);
-                }
-                return Ok(history);
-            }
-            (None, None) => {
-                let mut stmt = conn.prepare(
-                    "SELECT id, connection_id, connection_name, database_type, attempted_at, success, error_message, duration_ms
-                     FROM connection_history 
-                     ORDER BY attempted_at DESC 
-                     LIMIT ?1"
-                ).context("Failed to prepare connection history statement")?;
-
-                let rows = stmt
-                    .query_map([limit], |row| {
-                        Ok(ConnectionHistoryEntry {
-                            id: row.get(0)?,
-                            connection_id: row.get(1)?,
-                            connection_name: row.get(2)?,
-                            database_type: row.get(3)?,
-                            attempted_at: row.get(4)?,
-                            success: row.get(5)?,
-                            error_message: row.get(6)?,
-                            duration_ms: row.get(7)?,
-                        })
-                    })
-                    .context("Failed to query connection history")?;
-
-                let mut history = Vec::new();
-                for row in rows {
-                    history.push(row.context("Failed to process history row")?);
-                }
-                return Ok(history);
-            }
-        };
-    }
-
-    // =========================================================================
-    // Snippet Folder Methods
-    // =========================================================================
-
-    pub fn get_snippet_folders(&self) -> Result<Vec<SnippetFolder>> {
-        let conn = self.conn.lock().unwrap();
-        let mut stmt = conn.prepare(
-            "SELECT id, name, parent_id, color, created_at, updated_at 
-             FROM snippet_folders 
-             ORDER BY name ASC"
-        ).context("Failed to prepare snippet folders statement")?;
-
-        let rows = stmt
-            .query_map([], |row| {
-                Ok(SnippetFolder {
-                    id: row.get(0)?,
-                    name: row.get(1)?,
-                    parent_id: row.get(2)?,
-                    color: row.get(3)?,
-                    created_at: row.get(4)?,
-                    updated_at: row.get(5)?,
-                })
-            })
-            .context("Failed to query snippet folders")?;
-
-        let mut folders = Vec::new();
-        for row in rows {
-            folders.push(row.context("Failed to process folder row")?);
-        }
-        Ok(folders)
-    }
-
-    pub fn create_snippet_folder(&self, name: &str, parent_id: Option<i64>, color: Option<&str>) -> Result<i64> {
-        let now = chrono::Utc::now().timestamp();
-        let conn = self.conn.lock().unwrap();
-        
-        conn.execute(
-            "INSERT INTO snippet_folders (name, parent_id, color, created_at, updated_at) 
-             VALUES (?1, ?2, ?3, ?4, ?5)",
-            (name, parent_id, color, now, now),
-        ).context("Failed to create snippet folder")?;
-        
-        Ok(conn.last_insert_rowid())
-    }
-
-    pub fn update_snippet_folder(&self, id: i64, name: &str, color: Option<&str>) -> Result<()> {
-        let now = chrono::Utc::now().timestamp();
-        let conn = self.conn.lock().unwrap();
-        
-        conn.execute(
-            "UPDATE snippet_folders SET name = ?1, color = ?2, updated_at = ?3 WHERE id = ?4",
-            (name, color, now, id),
-        ).context("Failed to update snippet folder")?;
-        
-        Ok(())
-    }
-
-    pub fn delete_snippet_folder(&self, id: i64) -> Result<()> {
-        let conn = self.conn.lock().unwrap();
-        // First, set folder_id to NULL for all snippets in this folder
-        conn.execute(
-            "UPDATE saved_queries SET folder_id = NULL WHERE folder_id = ?1",
-            [id],
-        ).context("Failed to unlink snippets from folder")?;
-        
-        // Then delete the folder (cascade will handle child folders)
-        conn.execute("DELETE FROM snippet_folders WHERE id = ?1", [id])
-            .context("Failed to delete snippet folder")?;
-        
-        Ok(())
-    }
-}
